requirements.txt

Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
python-dotenv==1.0.0
